Use cache name in generate rule
This enables the building of multiple caches with the same configuration. Previously the configuration was used to define the generate rule's name which can lead to conflicts. The utility of this is low since embuilder already builds in parallel, but it prevents weird errors should the user try to attempt this.