Work in progress towards Bazel 8 support
This commit marks a pragmatic decision to wait some time before trying to support Bazel 8. The motiviation to support Bazel 8, mostly comes down to the ability to create a emscripten_binary macro. This macro would first delegate to cc_library to build a static library from all the provides sources before delegating to the internal rule which would run emcc to create the wasm, js, HTML output.

However, Bazel 8 (indirectly) required upgrading to a newer version of rules_cc which uses the new rule-based toolchain API. This has been time consuming to get working although some progress has been made. The current challenge in the code base is to reimplement the emscripten_merge_cache rule so that it returns a DirectoryInfo. The best approach I have to this is to copy the directory_rule implementation and modify it to merge the cache with everything that is needed.

The main reason for pausing this now is that it is time consuming and rules_cc is not quite ready (note that we are currently pinned to a specific commit since what I needed was not yet released). Furthermore, even if this was completely working, it is unclear whether or not rules_foreign_cc is compatible with Bazel 8 and the rule-based toolchain API.